+++ cpp/bench/sg/rf_regressor.cu
@@ -89,7 +89,7 @@ std::vector<RegParams> getInputs() {
                 500,   // n_trees
                 true,  // bootstrap
                 1.f,   // rows_sample
-                1234,  // seed
+                uint64_t(1234),  // seed
                 8);    // n_streams
 
   set_tree_params(p.rf.tree_params,  // Output tree parameters
